pyproject.toml
[project]
name = "git-tutor-bot"
version = "0.1.0"
description = "Git command explainer"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "litellm>=1.82.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
    "google-genai"
]
